Write Flash implemented

+++ src/isp.h
@@ -16,9 +16,9 @@ uint8_t ispEnterProgrammingMode(void);
 uint8_t ispReadEEPROM(unsigned int address);
 
 /* write byte to flash at given address */
-uint8_t ispWriteFlash(unsigned long address, uint8_t data, uint8_t pollmode);
+uint8_t ispWriteFlash(uint32_t address, uint8_t data, uint8_t pollmode);
 
-uint8_t ispFlushPage(unsigned long address, uint8_t pollvalue);
+uint8_t ispFlushPage(uint32_t address, uint8_t pollvalue);
 
 /* read byte from flash at given address */
 uint8_t ispReadFlash(uint32_t address);
